Macro rocket_okapi_codegen::routes_with_openapi

routes_with_openapi!() { /* proc-macro */ }

A replacement macro for rocket::routes. The key differences are that this macro will add an additional element to the resulting Vec<rocket::Route>, which serves a static file called openapi.json. This file can then be used to display the routes in the swagger ui.